feat: Add validation middleware for request body and ID

+++ src/index.js
@@ -6,44 +6,16 @@ import ffmpeg from 'fluent-ffmpeg';
 import { AssemblyAI } from "assemblyai";
 import { randomUUID } from 'node:crypto';
 
-import { Database } from './database.js';
 import createSummarization  from './queries/create-summarization.js';
 import getSummarization  from './queries/get-summarization.js';
+import { ValidateData } from "./class/validatedData.js";
 
 const server = Fastify();
-
-const database = new Database();
 
 const AssemblyAIClient = new AssemblyAI({
 	apiKey: process.env.ASSEMBLYAI_API_KEY
 });
 
-/**
- * Receber os dados do meu front-end [X]
- * validar os dados recebidos do meu front-end [X]
- * Transformar o video do link em um áudio e salvar ele temporiariamente na pasta public/audios
- * Enviar o audio para a API do assemblyai e pegar o job id daquela tarefa
- * salvar as informações so resumo no banco de dados local
- * Excluir o audio depois de criar a transcrição
- * 
- * 
- * Criar o banco de dados Postregres
- * Conectar com banco de dados
- * Salvar os dados no banco de dados
- */
-
-/**
- * Receber os dados do meu front-end
- * validar os dados recebidos do meu front-end
- * extrair o áudio do video e salvar localmente temporariamente na pasta public/audios
- * cortar o áudio para o tamanho definido
- * Enviar o áudio para API do assemblyai e pegar o job id daquela tarefa
- * Salvar o id do job no banco de dados
- * Verificar se o assemblyai já terminou o transcrever o áudio 
- * Pagar o áudio da pasta public
- */
-
-//created schema check for audio
 const audioSchema = z.object({
 	title: z.string(),
 	link: z.string().regex(new RegExp(/https:\/\/www\.youtube\.com\/watch\?v=[a-zA-Z0-9_-]{11}/)),
@@ -51,36 +23,76 @@ const audioSchema = z.object({
 	endAt: z.number(),
 });
 
+//verificar válidação do body
+const validateBody = (request, reply, done) => {
+	try {
+		if (request.method === 'POST' || request.method === 'PUT'){
+			if (Object.keys(request.body).length !==  4)
+				throw 'Body is invalid, check the body parameters.';
+			
+			const { title, link, startAt, endAt } = request.body;
+			
+			audioSchema.parse({title, link, startAt, endAt});
+			
+			done();
+		} else{
+			done();
+		}
+	} catch (error) {
+		return reply.status(500).send({ success: false, message: error });
+	}	
+}
+
+const validateId = (request, reply, done) => {
+	if (request.method === 'PUT' || request.method === 'DELETE'){
+		const { id } = request.params;
+
+		console.log(id);
+
+		const uuidv4Pattern = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+		if(!uuidv4Pattern.test(id)){
+			reply.code(400).send({ error: 'Invalid UUID v4 ID'});
+			return;
+		}
+		done();
+	}else{
+		done();
+	}
+}
+
+server.addHook('preHandler', validateBody);
+server.addHook('preHandler', validateId);
+
 server.post("/summarization", async (request, response) => {
 	try{
 		const { title, link, startAt, endAt } = request.body;
 		
-		const validateData = audioSchema.parse({title, link, startAt, endAt});
-
+		const validateData = new ValidateData(title, link, startAt, endAt);
+		
 		const videoReableStream = ytdl(validateData.link, {
 			quality: "lowestaudio"
 		});
-
+		
 		const filename = `${randomUUID()}.mp3`;
 		const outputFolder = "public/audios/";
-		
+
 		const ffmpegCommand = ffmpeg(videoReableStream)
-			.noVideo()
-			.audioCodec("libmp3lame")
-			.audioBitrate(128)
-			.seekInput(validateData.startAt)
-			.duration(validateData.endAt - validateData.startAt)
-			.format("mp3")
-			.output(`${outputFolder}${filename}`);
+		.noVideo()
+		.audioCodec("libmp3lame")
+		.audioBitrate(128)
+		.seekInput(validateData.startAt)
+		.duration(validateData.endAt - validateData.startAt)
+		.format("mp3")
+		.output(`${outputFolder}${filename}`);
+		
 		
 		await new Promise((resolve, reject) => {
 			ffmpegCommand.on("end", resolve);
 			ffmpegCommand.on("error", reject);
 			ffmpegCommand.run();
 		});
-	
-
-	/* 	const params = {
+		
+		/* const params = {
 			audio: `${outputFolder}${filename}`,
 			summarization: true,
 			summary_model: 'informative',
@@ -88,23 +100,14 @@ server.post("/summarization", async (request, response) => {
 		}
 
 		const transcript = await AssemblyAIClient.transcripts.transcribe(params);
-		
-		console.log(transcript);
-		
-			
- */
+		*/
+		validateData.updateTranscript("transcript.transcript");
+		validateData.updateSummary("transcript.summary");
 
+		if (!validateData.isValidInfo())
+			throw 'Information sent is invalid';
 
-		createSummarization(validateData);
-
-	/* 	database.create({
-			title: title,
-			link: link,
-			startAt, startAt,
-			endAt: endAt,
-			transcript: transcript.text,
-			summary: transcript.summary
-		}); */
+		await createSummarization(validateData);
 
 		unlink(`${outputFolder}${filename}`, (err) => {
 			if(err) return console.log(err);
@@ -119,9 +122,7 @@ server.post("/summarization", async (request, response) => {
 });
 
 server.get("/summarization", async (request, response) => {
-	
 	const summarization = await getSummarization();
-	console.log(summarization);
 	return response.status(200).send(summarization);
 });
 
@@ -142,7 +143,8 @@ server.put("/summarization/:id", (request, response) => {
 server.delete("/summarization/:id", (request, response) => {
 	const summarizationId = request.params.id;
 
-	database.delete(summarizationId);
+	console.log('Delete: ' + summarizationId);
+	//database.delete(summarizationId);
 
 	return response.status(200).send({"message" : "successfully removed"});
 });
@@ -150,4 +152,3 @@ server.delete("/summarization/:id", (request, response) => {
 server.listen({
 	port: 3333
 });
-
